Use partial classes for Graphics on Unix and Windows
